+++ app/src/main/java/com/saddict/rentalfinder/rentals/model/local/rentals/RentalEntity.kt
@@ -1,9 +1,10 @@
-package com.saddict.rentalfinder.rentals.model.local
+package com.saddict.rentalfinder.rentals.model.local.rentals
 
 import androidx.annotation.Keep
 import androidx.room.ColumnInfo
 import androidx.room.Entity
 import androidx.room.PrimaryKey
+import com.saddict.rentalfinder.rentals.model.remote.UserProfileDetails
 
 @Keep
 @Entity(tableName = "rentals")
@@ -17,7 +18,7 @@ data class RentalEntity(
     val title: String,
     val description: String,
     val category: String,
-    val location: String?,
+//    val location: String?,
     @ColumnInfo(name = "date_posted")
     val datePosted: String,
     @ColumnInfo(name = "time_modified")
@@ -31,12 +32,24 @@ data class RentalEntity(
     val isActive: Boolean,
     @ColumnInfo(name = "author_id")
     val authorId: Int,
-    @ColumnInfo(name = "author_first_name")
-    val authorFirstName: String?,
-    @ColumnInfo(name = "author_phone_number")
-    val authorPhoneNumber: String?,
     @ColumnInfo(name = "author_email")
     val authorEmail: String,
+    @ColumnInfo(name = "author_username")
+    val authorUsername: String,
+    @ColumnInfo(name = "author_profile_details")
+    val authorProfileDetails: UserProfileDetails,
+//    @ColumnInfo(name = "author_first_name")
+//    val authorFirstName: String?,
+//    @ColumnInfo(name = "author_last_name")
+//    val authorLastName: String?,
+//    @ColumnInfo(name = "author_phone_number")
+//    val authorPhoneNumber: String?,
+//    @ColumnInfo(name = "author_address")
+//    val authorAddress: String?,
+//    @ColumnInfo(name = "author_dob")
+//    val authorDob: String?,
+//    @ColumnInfo(name = "author_gender")
+//    val authorGender: String?,
     @ColumnInfo(name = "image_detail")
     val imageUrl: String,
     @ColumnInfo(name = "image_name")
