+++ app/src/main/java/com/saddict/rentalfinder/rentals/model/local/rentals/RentalRemoteKeysEntity.kt
@@ -1,4 +1,4 @@
-package com.saddict.rentalfinder.rentals.model.local
+package com.saddict.rentalfinder.rentals.model.local.rentals
 
 import androidx.annotation.Keep
 import androidx.room.ColumnInfo
@@ -6,8 +6,8 @@ import androidx.room.Entity
 import androidx.room.PrimaryKey
 
 @Keep
-@Entity(tableName = "image_remote_keys")
-data class ImageRemoteKeysEntity(
+@Entity(tableName = "rental_remote_keys")
+data class RentalRemoteKeysEntity(
     @PrimaryKey(autoGenerate = false)
     val id: Int,
     @ColumnInfo(name = "next_page")
