+++ app/src/main/java/com/saddict/rentalfinder/rentals/model/remote/rentals/Ratings.kt
@@ -1,4 +1,4 @@
-package com.saddict.rentalfinder.rentals.model.remote
+package com.saddict.rentalfinder.rentals.model.remote.rentals
 
 import androidx.annotation.Keep
 import com.fasterxml.jackson.annotation.JsonProperty
